+++ R/globals.R
@@ -818,5 +818,5 @@ utils::globalVariables(c(
 
 # .ng_breakdown() / .ng_play_type() (R/ng_breakdown.R).
 utils::globalVariables(c(
-  "value_own", "play_type", "net_goals", "ng_recon", "pre", "owed", "got", "tot", "s"
+  "value_own", "play_type", "net_goals", "ng_recon", "pre", "owed", "got", "tot", "s", "games", "n"
 ))